off-chip: add pac1934 driver to platform
These off-chip drivers should be included in the platform
repository. The reason being is that they aren't necessarily
third party, and when they are added to there are descrepencies
in the example projects because the power-saving examples have
these drivers but the other projects don't. This means that
the platforms aren't identical. Adding these off-chip drivers
to the platform solves this issue.

Jira-id: ESSBM-375
Signed-off-by: Charles Haden <Charles.Haden@microchip.com>