ddr: increase size of l2 lim cache flush
Customer was facing issues in trying to execute code
from the L2 lim memory. This was because the cache is
not being fully cleared for the executable code.

This size of the cache to be cleared in the ddr was
not the size it should have been and so it was patched
to increase the size by a multiple of 4, as seen in
the variable definitions.

Also converted some rebellious tabs into spaces.

Jira-id: ASKFPGA-3331
Jira-id: ESSBM-379
Signed-off-by: Charles Haden <Charles.Haden@microchip.com>